sbucket: sbucket_t -> sbucket, sbucket_idx_t -> string_idx_t, changed implementation
sbucket_t class has been renamed to sbucket. Only primitive types should have _t naming,
and sbucket_t is first to be changed according to this.

sbucket_idx_t was a misleading name, since the index isn't to the bucket, but to the
string itself. So it was renamed to string_idx_t.

The internal implementation wasn't obvious how it should work, so it's been redone and
comments have been added to make the design clear.

Also fixed a bug with string comparison so that it doesn't fail when it shouldn't.